+++ app/Domain/News/Services/NewsService.php
@@ -2,8 +2,12 @@
 
 namespace App\Domain\News\Services;
 
-use App\Domain\News\Entities\News;
 use App\Domain\News\Repositories\NewsRepositoryInterface;
+use App\Models\News;
+use Exception;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Facades\Log;
 
 class NewsService
 {
@@ -14,13 +18,24 @@ class NewsService
         $this->newsRepository = $newsRepository;
     }
 
-    public function getAll(): array
+    public function getAll(array $dataParams): LengthAwarePaginator
     {
-        return $this->newsRepository->getAll();
+        return $this->newsRepository->getAll($dataParams);
     }
 
-    public function getById(int $id): News
+    public function getById(int $id): ?News
     {
-        return $this->newsRepository->getById($id);
+        try {
+            $new = $this->newsRepository->getById($id);
+            if (empty($new)) {
+                throw new ModelNotFoundException('Noticia não foi encontrada ...');
+            }
+
+            return $new;
+        } catch (Exception $exception) {
+            Log::error($exception->getMessage());
+            throw new Exception($exception->getMessage());
+        }
+
     }
 }