+++ app/Infrastructure/Persistence/Eloquent/Repositories/EloquentNewsRepository.php
@@ -2,8 +2,10 @@
 
 namespace App\Infrastructure\Persistence\Eloquent\Repositories;
 
-use App\Domain\News\Entities\News;
 use App\Domain\News\Repositories\NewsRepositoryInterface;
+use App\Models\News;
+use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Facades\Cache;
 
 class EloquentNewsRepository implements NewsRepositoryInterface
 {
@@ -14,14 +16,22 @@ class EloquentNewsRepository implements NewsRepositoryInterface
         $this->newsModel = $newsModel;
     }
 
-    public function getAll(): array
+    public function getAll(array $dataParams): LengthAwarePaginator
     {
-        return $this->newsModel->paginate(10);
+        $serializetion = serialize($dataParams);
 
+        if (Cache::has('listNews'.$serializetion)) {
+            return Cache::get('listNews'.$serializetion);
+        }
+
+        return Cache::rememberForever('listNews'.$serializetion, function () use ($dataParams) {
+            return $this->newsModel->query()
+                ->paginate(perPage: $dataParams['per_page'] ?? 12, page: $dataParams['page'] ?? 1);
+        });
     }
 
-    public function getById(int $id): News
+    public function getById(int $id)
     {
-        return $this->newsModel->findOrFail($id);
+        return $this->newsModel->query()->find($id);
     }
 }