+++ database/migrations/2024_02_27_130551_create_news_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('news', function (Blueprint $table) {
             $table->id();
-            $table->string('title');
-            $table->string('description');
-            $table->string('full_txt');
+            $table->string('title', 256);
+            $table->string('description', 256);
+            $table->text('full_txt');
             $table->string('image');
             $table->date('publish_date');
             $table->timestamps();