+++ HISTORY.md
@@ -1,12 +1,41 @@
+## 1.1.1 (2024-02-29)
+
+* 
+
+## 1.1.1 (2024-02-29)
+
+* ajustando paginate e  per_page
+
+## 1.1.0 (2024-02-29)
+ 
+* adicionando cache
+
+## 1.0.9 (2024-02-28)
+
+* criando um seeder
+
+## 1.0.8 (2024-02-28)
+
+* criando um factory para news 
+* implementando factory
+
+## 1.0.7 (2024-02-27)
+
+* criando fron-end das noticias
+
+## 1.0.6 (2024-02-27)
+
+* criando rotas
+
 ## 1.0.5 (2024-02-27)
 
-
+* Implementando o controller
 
 ## 1.0.4 (2024-02-27)
 
 * criando EloquentNewsRepository que faz a implementaçao da NewsRepositoryInterface
 * instanciando os metodos da interface NewsRepositoryInterface
-* 
+
 ## 1.0.3 (2024-02-27)
 
 * criando service NewsService